+++ packages/valibot/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## next (YYYY-MM-DD)
 
 * Chore: update `@standard-schema/spec` to 1.0.0-beta.4
+* Chore: re-export `@gqloom/core`
 
 ## \[0.4.0] (2024-11-20)